Move CLI code to internal/app/singularity. Rename Driver to Bundle
